Refactor index.js to renderDOM.js and weatherAPI.js
-separated logic from dom-manipulation
